Added exception handling for the backend services

+++ backend/src/main/java/com/socialmedia/v/service/impl/ProfileServiceImpl.java
@@ -1,6 +1,9 @@
 package com.socialmedia.v.service.impl;
 
 import com.socialmedia.v.entity.Profile;
+import com.socialmedia.v.exception.BadRequestException;
+import com.socialmedia.v.exception.CustomException;
+import com.socialmedia.v.exception.NotFoundException;
 import com.socialmedia.v.repository.ProfileRepository;
 import com.socialmedia.v.service.ProfileService;
 import org.springframework.stereotype.Service;
@@ -12,8 +15,6 @@ import java.util.UUID;
 @Service
 public class ProfileServiceImpl implements ProfileService {
 
-
-
     private ProfileRepository profileRepository;
 
     public ProfileServiceImpl(ProfileRepository profileRepository) {
@@ -23,6 +24,14 @@ public class ProfileServiceImpl implements ProfileService {
     @Override
     public Profile create(Profile profile) {
 
+
+        if(profile.getUsername() ==null || profile.getUsername().isEmpty()){
+            throw new BadRequestException("company name should not be empty");
+        }
+
+        if(profile.getFullName() ==null || profile.getFullName().isEmpty()){
+            throw new BadRequestException("company name should not be empty");
+        }
 
 
         if(profileRepository.existsByUsername(profile.getUsername())){
@@ -36,7 +45,15 @@ public class ProfileServiceImpl implements ProfileService {
 
     @Override
     public List<Profile> getAll() {
-        return profileRepository.findAll();
+
+        List<Profile> profile = profileRepository.findAll();
+
+        if(profile ==null || profile.isEmpty()){
+            throw new CustomException("204","No Context","The List of profiles are Empty");
+        }
+
+
+        return profile;
     }
 
     @Override
@@ -44,18 +61,21 @@ public class ProfileServiceImpl implements ProfileService {
 
         Optional<Profile> profileOptional = profileRepository.findByUsername(username);
 
-        if(profileOptional.isPresent()){
+        if (profileOptional.isPresent()) {
             return profileOptional.get();
-        }else{
-            throw  new IllegalArgumentException("Profile Does not exists with given UserName: "+ username);
+        } else  {
+            throw new NotFoundException("Profile Does not exists with given UserName: " + username);
         }
-
     }
 
     @Override
     public Profile updateByUsername(String username, Profile profile) {
 
         Optional<Profile> profileOptional = profileRepository.findByUsername(username);
+
+        if(profileOptional.isEmpty() || profileOptional == null){
+            throw new NotFoundException(" No Profile Found with given username: "+ username);
+        }
 
         if(profileOptional.isPresent()){
             Profile profile1 = profileOptional.get();
@@ -66,9 +86,8 @@ public class ProfileServiceImpl implements ProfileService {
 
             return profileRepository.save(profile1);
         }else {
-            throw  new IllegalArgumentException("Profile with userName : "+ username + " Does not Exist");
+            throw  new RuntimeException("Something went Wrong");
         }
-
 
     }
 
@@ -77,12 +96,15 @@ public class ProfileServiceImpl implements ProfileService {
 
         Optional<Profile> profileOptional = profileRepository.findByUsername(username);
 
+        if (profileOptional.isEmpty() || profileOptional == null) {
+            throw new NotFoundException(" No Profile Found with given username: " + username);
+        }
 
-        if(profileOptional.isPresent()){
+        if (profileOptional.isPresent()) {
             profileRepository.delete(profileOptional.get());
 
-        }else {
-            throw new IllegalArgumentException("Profile with userName : " + username + "Does not exists to delete");
+        } else {
+            throw new RuntimeException("something went wrong");
         }
     }
 }
